% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bf_citation}
\alias{bf_citation}
\title{Cite all relevant items}
\usage{
bf_citation(registry)
}
\arguments{
\item{registry}{\code{\link[=registry]{registry(1)}}\cr a bitfield registry for which to
output all citations.}
}
\description{
Cite all relevant items
}
