% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_decimals.R
\name{bf_decimals}
\alias{bf_decimals}
\title{Build a bit flag by counting the number of decimals}
\usage{
bf_decimals(x, test, sep = "\\\\.", fill = TRUE)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} for which
decimals are checked.}

\item{sep}{\code{\link[=character]{character(1)}}\cr regex-compatible symbol that
separates the decimals from the numeric value.}

\item{fill}{\code{\link[=logical]{logical(1)}}\cr whether the function should consider
decimals lengths from 0 through the maximum length in \code{x[[test]]}, or
whether it should only document existing decimal lengths.}
}
\description{
Build a bit flag by counting the number of decimals
}
\details{
The show method of various classes shows decimals that may not
really be present, and that even includes ordinary numeric vectors. In
these cases, what is printed in the R console may be misleading.
Double check with \code{\link{str}}.

Determine the number of decimals of coordinates to determine their
precision:
\href{https://en.wikipedia.org/wiki/Decimal_degrees}{https://en.wikipedia.org/wiki/Decimal_degrees}
}
