% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_histogram.R
\name{bf_histogram}
\alias{bf_histogram}
\title{Build a bit flag from a histogram}
\usage{
bf_histogram(x)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains...}
}
\description{
A histogram a frequency distribution where the count of values is binned into
a small set of values so that it provides a simple overview over the values.
}
