% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_inf.R
\name{bf_inf}
\alias{bf_inf}
\title{Build a bit flag by checking that values are not Inf}
\usage{
bf_inf(x, test)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} that is
checked for Inf values.}
}
\description{
Build a bit flag by checking that values are not Inf
}
