% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bf_internal}
\alias{bf_internal}
\title{Internal operators}
\format{
a list containing encodings for the basic operators. Each operator is
a list with elements 'description', 'encoding_type', 'bits', 'test',
'require', 'extends', 'extends_note', 'version', and 'reference'.
}
\usage{
bf_internal
}
\description{
Internal operators
}
\keyword{datasets}
