% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_isCharacter.R
\name{bf_isCharacter}
\alias{bf_isCharacter}
\title{Build a bit flag by checking for a character value}
\usage{
bf_isCharacter(x, test, coerce = FALSE)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} for which a
character value is identified.}

\item{coerce}{\code{\link[=logical]{logical(1)}}\cr whether or not to carry out a more
aggressive test, where the values are coerced to the target type and those
that end up not having NA-values are reported as \code{TRUE}.}
}
\description{
Build a bit flag by checking for a character value
}
\details{
When coercing values to character, this function returns \code{TRUE}
for any value other than NA or Inf, and also for the all special cases of
NA (such as NA_character_, NA_real_, etc), as this is not supported in many
other programs that might provide data for or use them from this function.
}
