% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_isInteger.R
\name{bf_isInteger}
\alias{bf_isInteger}
\title{Build a bit flag by checking for a integer value}
\usage{
bf_isInteger(x, test, coerce = FALSE)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} for which a
integer value is identified.}

\item{coerce}{\code{\link[=logical]{logical(1)}}\cr whether or not to carry out a more
aggressive test, where the values are coerced to the target type and those
that end up not having NA-values are reported as \code{TRUE}.}
}
\description{
Build a bit flag by checking for a integer value
}
\details{
When coercing values to integer, they are typically truncated to the
non-decimal part of the value. This function compares the truncated value
to the original value and thus returns \code{TRUE} only when the original
value didn't have a decimals part or when that part was \code{*.0}. It,
moreover, returns \code{FALSE} for NA and NaN-values, and also for the
special case NA_integer_, as this is not supported in many other programs
that might provide data for or use them from this function.
}
