% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_isLogical.R
\name{bf_isLogical}
\alias{bf_isLogical}
\title{Build a bit flag by checking for a logical value}
\usage{
bf_isLogical(x, test, coerce = FALSE)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} for which a
logical value is identified.}

\item{coerce}{\code{\link[=logical]{logical(1)}}\cr whether or not to carry out a more
aggressive test, where the values are coerced to the target type and those
that end up not having NA-values are reported as \code{TRUE}.}
}
\description{
Build a bit flag by checking for a logical value
}
\details{
When coercing values to logical, this function returns \code{TRUE}
only for the integer (!) values \code{0} and \code{1}. For all other
values, including NA, this function returns \code{FALSE}.
}
