% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_isNumeric.R
\name{bf_isNumeric}
\alias{bf_isNumeric}
\title{Build a bit flag by checking for a numeric value}
\usage{
bf_isNumeric(x, test, coerce = FALSE)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} for which a
numeric value is identified.}

\item{coerce}{\code{\link[=logical]{logical(1)}}\cr whether or not to carry out a more
aggressive test, where the values are coerced to the target type and those
that end up not having NA-values are reported as \code{TRUE}.}
}
\description{
Build a bit flag by checking for a numeric value
}
\details{
When coercing values to numeric, this function returns \code{TRUE}
for any value that can sensibly be interpreted as numeric, such as
\code{1L} (an integer) or \code{"1"} (a character), but not \code{"1L"} (a
character of an integer) or \code{"one"} (another character). It returns
\code{FALSE} for NA and NaN-values, and also for the special case NA_real_,
as this is not supported in many other programs that might provide data for
or use them from this function.
}
