% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_length.R
\name{bf_length}
\alias{bf_length}
\title{Build a bit flag by counting the number of characters}
\usage{
bf_length(x, test)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} for which
decimals are checked.}
}
\description{
Build a bit flag by counting the number of characters
}
