% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_match.R
\name{bf_match}
\alias{bf_match}
\title{Build a bit flag by comparing a column with a set}
\usage{
bf_match(
  x,
  test,
  set,
  negate = FALSE,
  pos = NULL,
  na.val = NULL,
  description = NULL,
  registry = NULL
)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} for which a
match is checked.}

\item{set}{a vector of the same class as in \code{test} against which a set
operation is performed.}

\item{negate}{\code{\link[=logical]{logical(1)}}\cr whether or not to determine a subset
or a disjoint match.}

\item{pos}{\code{\link[=integer]{integerish(.)}}\cr the position(s) in the bitfield that
should be set.}

\item{na.val}{description}

\item{description}{description}

\item{registry}{description}
}
\description{
Build a bit flag by comparing a column with a set
}
\examples{
bf_match(x = tbl_bityield, test = "commodity", set = c("soybean", "maize"))

}
