% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_na.R
\name{bf_na}
\alias{bf_na}
\title{Build a bit by checking that values are not NA}
\usage{
bf_na(x, test)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} that is
checked for NA values.}
}
\description{
Build a bit by checking that values are not NA
}
