% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_nan.R
\name{bf_nan}
\alias{bf_nan}
\title{Build a bit flag by checking whether values are NaN}
\usage{
bf_nan(x, test, pos = NULL, na.val = NULL, description = NULL, registry = NULL)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} that is
checked for NaN values.}

\item{pos}{\code{\link[=integer]{integerish(.)}}\cr the position(s) in the bitfield that
should be set.}

\item{na.val}{description}

\item{description}{description}

\item{registry}{description}
}
\description{
Build a bit flag by checking whether values are NaN
}
\examples{
bf_nan(x = tbl_bityield, test = "y")
}
