% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_null.R
\name{bf_null}
\alias{bf_null}
\title{Build a bit flag by checking whether values are NULL}
\usage{
bf_null(x, test)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} that is
checked for NULL values.}
}
\description{
Build a bit flag by checking whether values are NULL
}
\examples{
bf_null(x = example_data, test = "z")
}
