% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_origin.R
\name{bf_origin}
\alias{bf_origin}
\title{Build a bit flag from an operation that has generated the focal entity}
\usage{
bf_origin(x)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains...}
}
\description{
Build a bit flag from an operation that has generated the focal entity
}
