% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bf_pcl}
\alias{bf_pcl}
\title{Internal bit-flag protocols}
\format{
a list containing bit-flag protocols for the internal tests. Each
protocol is a list itself with the fields \code{"name"}, \code{"version"},
\code{"extends"}, \code{"extends_note"}, \code{"description"},
\code{"encoding_type"}, \code{"bits"}, \code{"requires"}, \code{"test"},
\code{"data"} and \code{"reference"}. For information on how they were set
up and how you can set up additional protocols, go to
\code{\link{bf_protocol}}.
}
\usage{
bf_pcl
}
\description{
Internal bit-flag protocols
}
\keyword{datasets}
