% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_protocol.R
\name{bf_protocol}
\alias{bf_protocol}
\title{Define a new bit-flag protocol}
\usage{
bf_protocol(
  name,
  description,
  test,
  example,
  type,
  bits = NULL,
  extends = NULL,
  note = NULL,
  version = NULL,
  reference = NULL
)
}
\arguments{
\item{name}{description}

\item{description}{description}

\item{test}{\code{\link[=function]{function(...)}}\cr the function used to build the
bit flag.}

\item{example}{\code{\link[=list]{list(.)}}\cr named list that contains all arguments in
\code{test} as name with values of the correct type.}

\item{type}{description}

\item{bits}{description}

\item{extends}{description}

\item{note}{description}

\item{version}{description}

\item{reference}{description}
}
\description{
Define a new bit-flag protocol
}
\examples{
newFlag <- bf_protocol(name = "na",
                       description = "{x} contains NA-values{result}.",
                       test = function(x) is.na(x = x),
                       example = list(x = bf_tbl$commodity))
}
