% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_registry.R
\name{bf_registry}
\alias{bf_registry}
\title{Create a new registry}
\usage{
bf_registry(length = NULL, name = NULL, description = NULL)
}
\arguments{
\item{length}{\code{\link[=integer]{integerish(1)}}\cr how many observation there shall
be in the bitfield encoded by this registry.}

\item{name}{\code{\link[=character]{character(1)}}\cr the name of the bitfield.}

\item{description}{\code{\link[=character]{character(1)}}\cr the description of the
bitfield.}
}
\description{
Create a new registry
}
