% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_residuals.R
\name{bf_residuals}
\alias{bf_residuals}
\title{Build a bit flag from the residuals of a model}
\usage{
bf_residuals(x, resid)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains a column
with the residuals.}

\item{resid}{\code{\link[=character]{character(1)}}\cr the column in \code{x} that
contains the residuals.}
}
\description{
Build a bit flag from the residuals of a model
}
