% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_summarise.R
\name{bf_summarise}
\alias{bf_summarise}
\title{Build a bit flag from the summary of ...}
\usage{
bf_summarise(x, sum)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains a column
with the residuals.}

\item{sum}{\code{\link[=character]{character(1)}}\cr the column in \code{x} that ...}
}
\description{
Build a bit flag from the summary of ...
}
