% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_type.R
\name{bf_type}
\alias{bf_type}
\title{Build a bit flag by checking for the type}
\usage{
bf_type(x, test, type = NULL, coerce = FALSE)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} for which
the type is identified.}

\item{type}{\code{\link[=character]{character(1)}}\cr the for which to check, possible
values are \code{"integer"}, \code{"numeric"}, \code{"character"} or
\code{"logical"}.}

\item{coerce}{\code{\link[=logical]{logical(1)}}\cr whether or not to carry out a more
aggressive test, where the values are coerced to the target type and those
that end up not having NA-values are reported as \code{TRUE}.}
}
\value{
a logical vector of the same length as \code{test} with attributes
\code{name}, \code{description} and \code{triple}.
}
\description{
Build a bit flag by checking for the type
}
\details{
When coercing values to \itemize{ \item integer, they are typically
truncated to the non-decimal part of the value. This function compares the
truncated value to the original value and thus returns \code{TRUE} only
when the original value didn't have a decimals part or when that part was
\code{.0}. It, moreover, returns \code{FALSE} for \code{NA} and
\code{NaN}-values, and also for the special case \code{NA_integer_}, as
this is not supported in many other programs that might provide data for or
use them from this function. \item numeric, this function returns
\code{TRUE} for any value that can sensibly be interpreted as numeric, such
as \code{1L} (an integer) or \code{"1"} (a character), but not \code{"1L"}
(a character of an integer) or \code{"one"} (another character). It returns
\code{FALSE} for \code{NA} and \code{NaN}-values, and also for the special
case \code{NA_real_}, as this is not supported in many other programs that
might provide data for or use them from this function. \item character,
this function returns \code{FALSE} for Inf and \code{NA}, and also for the
special cases of \code{NA_character_}, as this is not supported in many
other programs that might provide data for or use them from this function.
\item logical, this function returns \code{TRUE} only for the integer(!)
values \code{0L} and \code{1L}. For all other values, including \code{NA},
this function returns \code{FALSE}. }
}
\examples{
bf_type(x = example_data, test = "y", type = "character", coerce = TRUE)
}
