% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1bitfield.R
\docType{class}
\name{bitfield-class}
\alias{bitfield-class}
\alias{bitfield}
\title{Bitfield class (S4) and methods}
\description{
A \code{bitfield} stores ...
}
\section{Slots}{

\describe{
\item{\code{width}}{\link{\code{integerish(1)}}\cr}

\item{\code{length}}{\link{\code{integerish(1)}}\cr}

\item{\code{name}}{\link{\code{character(1)}}\cr}

\item{\code{desc}}{\link{\code{data.frame(1)}}\cr}

\item{\code{bits}}{\link{\code{list(.)}}\cr}
}}

