% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bityield}
\alias{bityield}
\title{Example table}
\format{
object of class \code{tibble} has two columns that indicate
coordinates, one column that indicates a crop that is grown there, one
column that indicates the yield of that crop there and one column that
indicates the year of harvest. All columns contain some sort of deviation
that may occur in data.
}
\usage{
bityield
}
\description{
A 10 × 5 tibble with a range of example data to showcase functionality of
this package
}
\keyword{datasets}
