% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_precision.R
\name{calculate_precision}
\alias{calculate_precision}
\title{Calculate the precision from x and y coordinates}
\usage{
calculate_precision(input, area = FALSE)
}
\arguments{
\item{input}{the input dataframe that contains at least the columns x and y.}

\item{area}{whether or not to return also the corresponding area.}
}
\description{
Calculate the precision from x and y coordinates
}
