% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit.R
\name{.bit}
\alias{.bit}
\title{Derive bit representation}
\usage{
.bit(fun, flags = NULL)
}
\arguments{
\item{fun}{any function that returns either logical values or a sequence of
values starting at 0.}

\item{flags}{an integer value of the number of flags, can be higher than the
number of unique return values of fun, in case a specific instance of
\code{fun} doesn't cover the whole global standard.}
}
\description{
Derive bit representation
}
