% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.determineEncoding}
\alias{.determineEncoding}
\title{Make encoding for floating point values}
\usage{
.determineEncoding(
  x,
  precision = "single",
  decimals = NULL,
  range = NULL,
  fields = NULL
)
}
\arguments{
\item{x}{description}

\item{precision}{description}

\item{decimals}{\code{\link[=integer]{integer(1)}}\cr the number of decimal digits that should
be reliably represented, see Details.}

\item{range}{[\code{(1)}][]\cr the ratio between the smallest and largest possible value to be
reliably represented, see Details.}

\item{fields}{\code{\link[=list]{list(.)}}\cr options to control the encoding of the
binary representation of the numeric values.}
}
\description{
Make encoding for floating point values
}
\examples{
# example code

}
