% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.getBit}
\alias{.getBit}
\title{Extract bit values from integers}
\usage{
.getBit(x, len = NULL, rev = FALSE)
}
\arguments{
\item{x}{\code{\link[=integer]{integerish(1)}}\cr the integer from which to extract a
bit.}

\item{len}{\code{\link[=integer]{integerish(1)}}\cr the length of the returned bit.}

\item{rev}{\code{\link[=logical]{logical(1)}}\cr whether or not to revert the direction
of how the bit is returned.}
}
\description{
Extract bit values from integers
}
