% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.getDependencies}
\alias{.getDependencies}
\title{Identify packages to custom functions}
\usage{
.getDependencies(fun)
}
\arguments{
\item{fun}{\code{\link[=function]{function(...)}}\cr the custom function in which to
identify dependencies.}
}
\value{
vector of packages that are required to run the function.
}
\description{
Identify packages to custom functions
}
