% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBits.R
\name{.intToBin}
\alias{.intToBin}
\title{Make a bit flag from an integer}
\usage{
.intToBin(x, len = NULL)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric(1)}}\cr the numeric value for which to derive the
binary value.}

\item{len}{\code{\link[=integer]{integerish(1)}}\cr the number of bits used to capture
the value.}
}
\description{
Make a bit flag from an integer
}
\details{
Additional details...
}
\examples{
# example code

}
