% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBits.R
\name{.makeBits}
\alias{.makeBits}
\title{Make a bit flag from an integer}
\usage{
.makeBits(x, encoding = NULL)
}
\arguments{
\item{x}{the integer for which to derive the
binary representation (bit flags).}

\item{encoding}{the number of bit flags.}
}
\description{
Make a bit flag from an integer
}
