% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.makeDatacite}
\alias{.makeDatacite}
\title{Create DataCite-compliant metadata structure}
\usage{
.makeDatacite(registry)
}
\arguments{
\item{registry}{Registry object}
}
\value{
List with DataCite-compliant structure
}
\description{
Create DataCite-compliant metadata structure
}
