% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.rast}
\alias{.rast}
\title{Extract values and metadata from terra::SpatRaster}
\usage{
.rast(x)
}
\arguments{
\item{x}{\link[terra:SpatRaster-class]{SpatRaster(1)}\cr the SpatRaster object.}
}
\value{
the function that extracts values and metadata.
}
\description{
Extract values and metadata from terra::SpatRaster
}
\details{
This function simply extracts the values from \code{x} and appends
the raster metadata as attributes.
}
