% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.toBin}
\alias{.toBin}
\title{Make a binary value from an integer}
\usage{
.toBin(x, len = NULL, pad = TRUE)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric(.)}}\cr numeric vector for which to derive the
binary values.}

\item{len}{\code{\link[=integer]{integerish(1)}}\cr the number of bits used to capture
each value. If NULL, computed from the maximum value.}

\item{pad}{\code{\link[=logical]{logical(1)}}\cr whether to pad the binary values with
leading zeros to equal width.}
}
\description{
Make a binary value from an integer
}
