% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.toDec}
\alias{.toDec}
\title{Make an integer from a binary value}
\usage{
.toDec(x)
}
\arguments{
\item{x}{\code{\link[=character]{character(1)}}\cr a binary string (sequence of 0s and
1s) for which to derive the integer.}
}
\description{
Make an integer from a binary value
}
