% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.toEncoding}
\alias{.toEncoding}
\title{Make encoding for floating point values}
\usage{
.toEncoding(x, precision = NULL, range = NULL, opts = NULL)
}
\arguments{
\item{x}{description}

\item{precision}{\code{\link[=integer]{integer(1)}}\cr the number of digits that should
be reliably represented, see Details.}

\item{range}{[\code{(1)}][]\cr the ratio between the smallest and largest possible value to be
reliably represented, see Details.}

\item{opts}{\code{\link[=list]{list(.)}}\cr options to control the encoding of the
binary representation of the numeric values, see Details.}
}
\description{
Make encoding for floating point values
}
\details{
\code{precision} modifies the significand \code{range} modifies the
exponent
}
\examples{
# example code

}
