% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.toYml}
\alias{.toYml}
\title{Express a bit-flag protocol as yaml file}
\usage{
.toYml(pcl, file)
}
\arguments{
\item{pcl}{description}

\item{file}{description}
}
\value{
called for the side effect of storing a yml file.
}
\description{
Express a bit-flag protocol as yaml file
}
