% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.updateMD5}
\alias{.updateMD5}
\title{Determine and write MD5 sum}
\usage{
.updateMD5(x)
}
\arguments{
\item{x}{\code{\link[=registry]{registry(1)}}\cr registry for which to determine the MD5
checksum.}
}
\value{
this function is called for its side-effect of storing the MD5
checksum in the md5 slot of the registry.
}
\description{
Determine and write MD5 sum
}
\details{
This function follows the following algorithm: \itemize{ \item set
the current MD5 checksum to NA_character_, \item write the registry into
the temporary directory, \item calculate the checksum of this file and
finally \item store the checksum  in the md5 slot of the registry.} This
means that when comparing the MD5 checksum in this slot, one first has to
set that value also to NULL, otherwise the two values won't coincide.
}
