% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.validate_token}
\alias{.validate_token}
\title{Validate a github token}
\usage{
.validate_token(token)
}
\arguments{
\item{token}{\code{\link[=character]{character(1)}}\cr github PAT (personal access
token).}
}
\value{
the validated user token
}
\description{
This function checks whether the user-provided token is valid for use with
this package.
}
