% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.validateProtocol}
\alias{.validateProtocol}
\title{Validate a bit-flag protocol}
\usage{
.validateProtocol(protocol)
}
\arguments{
\item{protocol}{the protocol to validate}
}
\value{
the validated protocol
}
\description{
Validate a bit-flag protocol
}
