% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_analyze.R
\name{print.bf_analysis}
\alias{print.bf_analysis}
\title{Print method for bf_analysis}
\usage{
\method{print}{bf_analysis}(x, min_bits = NULL, ...)
}
\arguments{
\item{x}{bf_analysis object}

\item{min_bits}{minimum total bits to display (overrides value from
bf_analyze if provided)}

\item{...}{additional arguments (ignored)}
}
\description{
Print method for bf_analysis
}
