% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1protocol.R
\docType{class}
\name{protocol-class}
\alias{protocol-class}
\alias{protocol}
\title{Bitflag class (S4) and methods}
\description{
A \code{protocol} stores test, documentation and example data to establish a
formal method to build a bit flag.
}
\section{Slots}{

\describe{
\item{\code{meta}}{description}

\item{\code{test}}{description}

\item{\code{specs}}{description}

\item{\code{extends}}{description}
}}

