% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_case.R
\name{qb_case}
\alias{qb_case}
\title{Build a bit by checking}
\usage{
qb_case(x, ..., exclusive = TRUE)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{...}{any statement that results in a logical return value}
}
\description{
Build a bit by checking
}
