% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_combine.R
\name{qb_combine}
\alias{qb_combine}
\title{Combine the bitfield into a quality bit (QB)}
\usage{
qb_combine(bitfield)
}
\arguments{
\item{bitfield}{\code{\link[=bitfield]{bitfield(1)}}\cr the bitfield that should be
combined into a quality bit (QB).}
}
\description{
Combine the bitfield into a quality bit (QB)
}
