% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_grow.R
\name{qb_grow}
\alias{qb_grow}
\title{Add bits to a bitfield}
\usage{
qb_grow(bit, desc, na = FALSE, pos, bitfield)
}
\arguments{
\item{bit}{a function that returns}

\item{desc}{the description of the bit(s).}

\item{na}{which value NAs in the output of .bit should have in the bitfield.}

\item{pos}{the position in the bitfield that should be set.}

\item{bitfield}{the bitfield in which the bits should be set.}
}
\description{
Add bits to a bitfield
}
