% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_harvest.R
\name{qb_harvest}
\alias{qb_harvest}
\title{Combine the bitfield into a quality bit (QB)}
\usage{
qb_harvest(bitfield)
}
\arguments{
\item{bitfield}{\code{\link[=bitfield]{bitfield(1)}}\cr the bitfield that should be
combined into a quality bit (QB).}
}
\description{
Combine the bitfield into a quality bit (QB)
}
