% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_precision.R
\name{qb_precision}
\alias{qb_precision}
\title{Build a bit by checking}
\usage{
qb_precision(x, test)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} that is
checked for NA values.}
}
\description{
Build a bit by checking
}
