% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_unpack.R
\name{qb_unpack}
\alias{qb_unpack}
\title{Unpack the integer representation of QB}
\usage{
qb_unpack(x, bitfield, sep = ".")
}
\arguments{
\item{x}{description}

\item{bitfield}{\code{\link[=bitfield]{bitfield(1)}}\cr the bitfield that should be
unpacked into a quality bit (QB).}

\item{sep}{the value by which the distinct fields shall be separated}
}
\description{
Unpack the integer representation of QB
}
