% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbb_filter.R
\name{qbb_filter}
\alias{qbb_filter}
\title{Build a bit by checking for a match with an external vector (not working yet)}
\usage{
qbb_filter(x, test, condition)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that
contains \code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in
\code{x} for which a match is checked.}

\item{condition}{\cr any statement that results in a logical return value}
}
\description{
Build a bit by checking for a match with an external vector (not working yet)
}
