% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbb_match.R
\name{qbb_match}
\alias{qbb_match}
\title{Build a bit by checking for a match with an external vector}
\usage{
qbb_match(x, test, against)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that
contains \code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in
\code{x} for which a match is checked.}

\item{against}{\code{\link[=numeric]{numeric(.)}} | \code{\link[=character]{character(.)}} |
\code{\link[=logical]{logical(1)}}\cr a vector of the same class as in \code{test}
against which a match is performed.}
}
\description{
Build a bit by checking for a match with an external vector
}
