% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbb_na.R
\name{qbb_na}
\alias{qbb_na}
\title{Build a bit by checking for NA values}
\usage{
qbb_na(x, test)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} that is
checked for NA values.}
}
\description{
Build a bit by checking for NA values
}
