% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbb_range.R
\name{qbb_range}
\alias{qbb_range}
\title{Build a bit by checking for a range}
\usage{
qbb_range(x, test, min, max)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} for which a
range is checked.}

\item{min}{\code{\link[=numeric]{numeric(1)}}\cr the minimum allowed value.}

\item{max}{\code{\link[=numeric]{numeric(1)}}\cr the maximum allowed value.}
}
\description{
Build a bit by checking for a range
}
