% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1registry.R
\docType{class}
\name{registry-class}
\alias{registry-class}
\alias{registry}
\title{Bit registry class (S4) and methods}
\description{
A \code{registry} stores ...
}
\section{Slots}{

\describe{
\item{\code{width}}{\code{\link[=integer]{integerish(1)}}\cr}

\item{\code{length}}{\code{\link[=integer]{integerish(1)}}\cr}

\item{\code{name}}{\code{\link[=character]{character(.)}}\cr}

\item{\code{desc}}{\code{\link[=data.frame]{data.frame(1)}}\cr}

\item{\code{bits}}{\code{\link[=list]{list(.)}}\cr}
}}

