% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1protocol.R
\name{show,protocol-method}
\alias{show,protocol-method}
\title{Print protocol in the console}
\usage{
\S4method{show}{protocol}(object)
}
\arguments{
\item{object}{\code{\link[=protocol]{protocol(1)}}\cr object to \code{show}.}
}
\description{
Print protocol in the console
}
