% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1registry.R
\name{show,registry-method}
\alias{show,registry-method}
\title{Print registry in the console}
\usage{
\S4method{show}{registry}(object)
}
\arguments{
\item{object}{\code{\link[=registry]{registry(1)}}\cr object to \code{show}.}
}
\description{
Print registry in the console
}
\details{
This method produces an overview of the registry by printing a
header with information about the setup of the bitfield and a table with
one line for each flag in the bitfield. The table shows the start position
of each flag, the encoding type (see \code{\link{.makeEncoding}}), the
bitfield operator type and the columns that are tested by the flag.
}
