% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCoords.R
\name{testCoords}
\alias{testCoords}
\title{Test coordinates for plausibility}
\usage{
testCoords(input)
}
\arguments{
\item{input}{\code{\link{data.frame}}\cr tidy table that contains
coordinate values in the columns 'x' and 'y'.}
}
\description{
Test coordinates for plausibility
}
