% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_apply.R
\name{bf_apply}
\alias{bf_apply}
\title{Build a bit flag by applying a standard encoding}
\usage{
bf_apply(
  x,
  test,
  standard,
  pos = NULL,
  na.val = NULL,
  description = NULL,
  registry = NULL
)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the object to build bit flags for.}
}
\description{
Build a bit flag by applying a standard encoding
}
