% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_create.R
\name{bf_create}
\alias{bf_create}
\title{Create a new bitfield}
\usage{
bf_create(width, length, name = NULL)
}
\arguments{
\item{width}{\code{\link[=integer]{integerish(1)}}\cr the table that contains
\code{test}.}

\item{length}{\code{\link[=integer]{integerish(1)}}\cr the column in \code{x} for which
a range is checked.}

\item{name}{\code{\link[=character]{character(1)}}\cr the minimum allowed value.}
}
\description{
Create a new bitfield
}
