% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_distribution.R
\name{bf_distribution}
\alias{bf_distribution}
\title{Build a bit flag from a (probability density) distribution}
\usage{
bf_distribution(x, ..., familiy = NULL)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{source}.}

\item{...}{description}

\item{family}{\code{\link[=character]{character(1)}}\cr the family of the density
distribution.}
}
\description{
Build a bit flag from a (probability density) distribution
}
