% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_encode.R
\name{bf_encode}
\alias{bf_encode}
\title{Build a bitfield from a bit registry}
\usage{
bf_encode(registry)
}
\arguments{
\item{registry}{\code{\link[=registry]{registry(1)}}\cr the bit registry that should be
combined into a bitfield.}
}
\description{
Build a bitfield from a bit registry
}
