% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_encode.R
\name{bf_encode}
\alias{bf_encode}
\title{Encode bit flags into a bitfield}
\usage{
bf_encode(registry)
}
\arguments{
\item{registry}{\code{\link[=registry]{registry(1)}}\cr the registry that should be
encoded into a bitfield.}
}
\value{
data.frame of the same length as the input data. Depending on type
and amount of bit flags, this can a table with any number of columns, each
of which encodes a sequence of 32 bits into an integer.
}
\description{
This function picks up the flags mentioned in a registry and encodes them as
integer values.
}
\examples{
reg <- bf_map(protocol = "na", data = bf_tbl, x = y)

field <- bf_encode(registry = reg)
}
