% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_encode.R
\name{bf_encode}
\alias{bf_encode}
\title{Build a bitfield from a bit registry}
\usage{
bf_encode(registry)
}
\arguments{
\item{registry}{\code{\link[=registry]{registry(1)}}\cr the bit registry that should be
encoded into a bitfield.}
}
\value{
description
}
\description{
Build a bitfield from a bit registry
}
\examples{
myRegistry <- bf_na(bf_tbl, "y")

myBitfield <- bf_encode(registry = myRegistry)
}
