% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_export.R
\name{bf_export}
\alias{bf_export}
\title{Export bitfield registries}
\usage{
bf_export(registry, format = NULL, file = NULL)
}
\arguments{
\item{registry}{description}

\item{format}{description}

\item{file}{description}
}
\value{
description
}
\description{
Export bitfield registries
}
\examples{
bf_export(reg, format = "json", file = "yield_QA.json")
bf_export(reg, format = "xml", file = "yield_QA.xml")
bf_export(reg, format = "yaml", file = "yield_QA.yaml")
}
