% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_identical.R
\name{bf_identical}
\alias{bf_identical}
\title{Build a bit flag by checking whether two columns are identical}
\usage{
bf_identical(x, test, against)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test} and \code{against}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} that is
checked against \code{against}.}

\item{against}{\code{\link[=character]{character(1)}}\cr the column in \code{x} that is
checked against \code{test}.}
}
\description{
Build a bit flag by checking whether two columns are identical
}
\details{
This function compares the values of two columns element-wise and
returns \code{TRUE} when they are identical.
}
\examples{
bf_identical(x = example_data, test = "x", against = "y")
}
