% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_match.R
\name{bf_match}
\alias{bf_match}
\title{Build a bit flag by comparing a column with a set}
\usage{
bf_match(
  x,
  test,
  set,
  negate = FALSE,
  pos = NULL,
  na.val = NULL,
  description = NULL,
  registry = NULL
)
}
\arguments{
\item{x}{the object to build bit flags for.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column or layer in \code{x}
for which a match is checked.}

\item{set}{a vector of the same class as in \code{test} against which a set
operation is performed.}

\item{negate}{\code{\link[=logical]{logical(1)}}\cr whether or not to determine a subset
or a disjoint match.}

\item{pos}{\code{\link[=integer]{integerish(.)}}\cr the position(s) in the bitfield that
should be set.}

\item{na.val}{\code{\link[=character]{character(1)}}\cr optional value that should be
used to substitute NA values in the input data.}

\item{description}{\code{\link[=character]{character(.)}}\cr optional description that
should be used instead of the default function-specific description. This
description is used in the registry legend, so it should have as many
entries as there will be entries per the respective flag in the legend (two
for a binary flag, as many as there are cases for a cases flag and one for
count or numeric flags).}

\item{registry}{\code{\link[=registry]{registry(1)}}\cr a bitfield registry that has
been defined with \code{\link{bf_registry}}; if it's undefined, an empty
registry will be defined on-the-fly.}
}
\value{
an (updated) object of class 'registry' with the additional flag
defined here.
}
\description{
Build a bit flag by comparing a column with a set
}
\examples{
bf_match(x = tbl_bityield, test = "commodity", set = c("soybean", "maize"))
}
