% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_na.R
\name{bf_na}
\alias{bf_na}
\title{Build a bit flag by checking whether values are NA}
\usage{
bf_na(x, test)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} that is
checked for NA values.}
}
\description{
Build a bit flag by checking whether values are NA
}
\examples{
bf_na(x = example_data, test = "y")
}
