% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_nan.R
\name{bf_nan}
\alias{bf_nan}
\title{Build a bit flag by checking that values are not NaN}
\usage{
bf_nan(x, test)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} that is
checked for NaN values.}
}
\description{
Build a bit flag by checking that values are not NaN
}
