% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_nan.R
\name{bf_nan}
\alias{bf_nan}
\title{Build a bit flag by checking whether values are NaN}
\usage{
bf_nan(x, test, pos = NULL, na.val = NULL, description = NULL, registry = NULL)
}
\arguments{
\item{x}{the object to build bit flags for.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column or layer in \code{x}
that is checked for NaN values.}

\item{pos}{\code{\link[=integer]{integerish(.)}}\cr the position(s) in the bitfield that
should be set.}

\item{na.val}{\code{\link[=logical]{logical(1)}}\cr value that needs to be given, if the
test for this flag results in \code{NA}s.}

\item{description}{\code{\link[=character]{character(.)}}\cr optional description that
should be used instead of the default function-specific description. This
description is used in the registry legend, so it should have as many
entries as there will be entries per the respective flag in the legend (two
for a binary flag, as many as there are cases for a cases flag and one for
count or numeric flags).}

\item{registry}{\code{\link[=registry]{registry(1)}}\cr a bitfield registry that has
been defined with \code{\link{bf_registry}}; if it's undefined, an empty
registry will be defined on-the-fly.}
}
\value{
an (updated) object of class 'registry' with the additional flag
defined here.
}
\description{
Build a bit flag by checking whether values are NaN
}
\examples{
bf_nan(x = bf_tbl, test = "y")
}
