% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_numeric.R
\name{bf_numeric}
\alias{bf_numeric}
\title{Build a bit flag from a numeric value}
\usage{
bf_numeric(
  x,
  source,
  ...,
  pos = NULL,
  na.val = NULL,
  description = NULL,
  registry = NULL
)
}
\arguments{
\item{x}{the object to build bit flags for.}

\item{source}{\code{\link[=character]{character(1)}}\cr the column or layer in \code{x}
from which to take the numeric value.}

\item{...}{description}

\item{pos}{\code{\link[=integer]{integerish(.)}}\cr the position(s) in the bitfield that
should be set.}

\item{na.val}{\code{\link[=logical]{logical(1)}}\cr value that needs to be given, if the
test for this flag results in \code{NA}s.}

\item{description}{\code{\link[=character]{character(.)}}\cr optional description that
should be used instead of the default function-specific description. This
description is used in the registry legend, so it should have as many
entries as there will be entries per the respective flag in the legend (two
for a binary flag, as many as there are cases for a cases flag and one for
count or numeric flags).}

\item{registry}{\code{\link[=registry]{registry(1)}}\cr a bitfield registry that has
been defined with \code{\link{bf_registry}}; if it's undefined, an empty
registry will be defined on-the-fly.}
}
\value{
an (updated) object of class 'registry' with the additional flag
defined here.
}
\description{
Build a bit flag from a numeric value
}
\details{
The length of the bitfield depends on the floating point precision
of the numeric values returned with this function. This is given as a
sequence of 3 integers of the form [1.5.10]. The first position encodes
how many bit are used (0 or 1), the second position encodes how many
exponent bits are used and the third position encodes how many significant
fields are used. See \code{\link{.determineEncoding}} for details.
}
\examples{
bf_numeric(x = bf_tbl, source = "yield")
}
