% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_numeric.R
\name{bf_numeric}
\alias{bf_numeric}
\title{Build a bit flag from a numeric value}
\usage{
bf_numeric(x, source, digits = 0)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{source}.}

\item{source}{\code{\link[=character]{character(1)}}\cr the column in \code{x} from
which to take the numeric value.}

\item{digits}{\code{\link[=integer]{integerish(1)}}\cr the number of decimals to which
to round the numeric value.}
}
\description{
Build a bit flag from a numeric value
}
\details{
The length of the bitfield depends on the floating point precision
of the numeric values returned with this function...
}
\examples{
bf_numeric(x = example_data, source = "y", digits = 0)
}
