% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bf_pcl}
\alias{bf_pcl}
\title{Internal bitflag protocols}
\format{
a list containing bitflag protocols for the internal tests. Each element is
an item of class \code{\link{protocol}}.
}
\usage{
bf_pcl
}
\description{
Internal bitflag protocols
}
\keyword{datasets}
