% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_range.R
\name{bf_range}
\alias{bf_range}
\title{Build a bit flag by checking for a range}
\usage{
bf_range(
  x,
  test,
  min = NULL,
  max = NULL,
  pos = NULL,
  na.val = NULL,
  description = NULL,
  registry = NULL
)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame(1)}}\cr the table that contains
\code{test}.}

\item{test}{\code{\link[=character]{character(1)}}\cr the column in \code{x} for which a
range is checked.}

\item{min}{\code{\link[=numeric]{numeric(1)}}\cr the minimum allowed value.}

\item{max}{\code{\link[=numeric]{numeric(1)}}\cr the maximum allowed value.}

\item{pos}{\code{\link[=integer]{integerish(.)}}\cr the position(s) in the bitfield that
should be set.}

\item{na.val}{\code{\link[=character]{character(1)}}\cr optional value that should be
used to substitute NA values in the input data.}

\item{description}{\code{\link[=character]{character(.)}}\cr optional description that
should be used instead of the default function-specific description. This
description is used in the registry legend, so it should have as many
entries as there will be entries per the respective flag in the legend (two
for a binary flag, as many as there are cases for a cases flag and one for
count or numeric flags).}

\item{registry}{\code{\link[=registry]{registry(1)}}\cr a bitfield registry that has
been defined with \code{\link{bf_registry}}; if it's undefined, an empty
registry will be defined on-the-fly.}
}
\value{
an object of class 'registry' with the additional flag defined here.
}
\description{
Build a bit flag by checking for a range
}
\details{
when leaving either \code{min} or \code{max} at NULL they are set to
the minimum or maximum value of \code{test}, respectively, thereby carrying
out a "less than max" or "greater than min" operation.
}
\examples{
bf_range(x = tbl_bityield, test = "y", min = 50, max = 59, na.val = 0)
}
