% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_registry.R
\name{bf_registry}
\alias{bf_registry}
\title{Create a new registry}
\usage{
bf_registry(name = NULL, description = NULL)
}
\arguments{
\item{name}{\code{\link[=character]{character(1)}}\cr the name of the bitfield.}

\item{description}{\code{\link[=character]{character(1)}}\cr the description of the
bitfield.}
}
\value{
an empty registry that captures some metadata of the bitfield, but
doesn't contain any flags yet.
}
\description{
Create a new registry
}
\details{
Besides creating a registry object, this function also opens the
environment \code{bf_env}, which is used to store the temporary non-encoded
flag values.
}
\examples{
reg <- bf_registry(name = "currentWorkflow",
                   description = "this is to document my current workflow so
                                  that I can share it with my colleagues
                                  alongside a publication.")
}
