% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_unpack.R
\name{bf_unpack}
\alias{bf_unpack}
\title{Unpack a bitfield}
\usage{
bf_unpack(x, registry, merge = NULL)
}
\arguments{
\item{x}{\code{\link[=integer]{integerish(1)}}\cr table of the integer representation of
the bitfield.}

\item{registry}{\code{\link[=registry]{registry(1)}}\cr the registry that should be
used to unpack the bitfield into a binary representation.}

\item{merge}{\code{\link[=character]{character(1)}}\cr a symbol with which, if given,
the distinct fields shall be merged.}
}
\description{
Unpack a bitfield
}
