% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1bitfield.R
\docType{class}
\name{bitfield-class}
\alias{bitfield-class}
\alias{bitfield}
\title{Bitfield class (S4) and methods}
\description{
A \code{bitfield} stores ...
}
\section{Slots}{

\describe{
\item{\code{width}}{\code{\link[=integer]{integerish(1)}}\cr}

\item{\code{length}}{\code{\link[=integer]{integerish(1)}}\cr}

\item{\code{name}}{\code{\link[=character]{character(.)}}\cr}

\item{\code{desc}}{\code{\link[=data.frame]{data.frame(.)}}\cr}

\item{\code{bits}}{\code{\link[=list]{list()}}\cr}
}}

