% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfield-package.R
\docType{package}
\name{bitfield-package}
\alias{bitfield}
\alias{bitfield-package}
\title{bitfield: Handle Bitfields to record Data-Quality}
\description{
The bitfield package provides tools to record meta-analytic and
meta-algorithmic data or just ordinary values to store those in a bitfield. A
bitfield accompanies a (modelled) dataset and can give insight into data
quality, provenance, and intermediate values, or can be used to store output
values in a highly compressed form.
}
\details{
wip
}
\seealso{
\itemize{ \item Github project:
\href{https://github.com/EhrmannS/bitfield}{https://github.com/EhrmannS/bitfield}
\item Report bugs:
\href{https://github.com/EhrmannS/bitfield/issues}{https://github.com/EhrmannS/bitfield/issues}
}
}
\author{
\strong{Maintainer, Author}: Steffen Ehrmann
\email{steffen.ehrmann@posteo.de}
}
