% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfield-package.R
\docType{package}
\name{bitfield-package}
\alias{bitfield}
\alias{bitfield-package}
\title{bitfield: Handle Bitfields to record Meta Data}
\description{
The bitfield package provides tools to record analytic and algorithmic meta
data or just any ordinary values to store in a bitfield. A bitfield
accompanies a (modelled) dataset and can give insight into data quality,
provenance, and intermediate values, or can be used to store output values in
a highly compressed form.
}
\details{
The general workflow consists of defining a registry, growing the bitfield by
specifying tests via the operators, to encode this into an integer value that
can be stored and published, or decoded and re-used in a downstream
application.
}
\seealso{
\itemize{ \item Github project:
\href{https://github.com/bitfloat/bitfield}{https://github.com/bitfloat/bitfield}
\item Report bugs:
\href{https://github.com/bitfloat/bitfield/issues}{https://github.com/bitfloat/bitfield/issues}
}
}
\author{
\strong{Maintainer, Author}: Steffen Ehrmann
\email{steffen.ehrmann@posteo.de}
}
