% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBits.R
\name{.makeBits}
\alias{.makeBits}
\title{Make a bit flag from an integer}
\usage{
.makeBits(x, encoding = NULL)
}
\arguments{
\item{x}{\code{\link[=integer]{integerish(1)}}\cr the integer for which to derive the
binary representation (bit flags).}

\item{len}{\code{\link[=integer]{integerish(1)}}\cr the number of bit flags.}

\item{rev}{\code{\link[=logical]{logical(1)}}\cr whether or not to revert the direction
of how the bit flag is returned.}
}
\description{
Make a bit flag from an integer
}
\details{
Wrapper around \code{\link{intToBits}}.
}
