% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.toBin}
\alias{.toBin}
\title{Make a binary value from an integer}
\usage{
.toBin(x, len = NULL, pad = TRUE, dec = FALSE)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric(1)}}\cr the numeric value for which to derive the
binary value.}

\item{len}{\code{\link[=integer]{integerish(1)}}\cr the number of bits used to capture
the value.}

\item{pad}{\code{\link[=logical]{logical(1)}}\cr whether to pad the binary value with 0
values.}

\item{dec}{\code{\link[=logical]{logical(1)}}\cr whether to transform the decimal part
to bits, or the integer part.}
}
\description{
Make a binary value from an integer
}
\details{
Additional details...
}
\examples{
# example code

}
