% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_combine.R
\name{qb_combine}
\alias{qb_combine}
\title{Add new bits to a bitfield}
\usage{
qb_combine(bitfield, inspect = FALSE)
}
\arguments{
\item{bitfield}{\code{\link[=bitfield]{bitfield(1)}}\cr the bitfield with which the bits
should be combined.}

\item{inspect}{\code{\link[=logical]{logical(1)}}\cr whether or not to inspect the bit.}
}
\description{
Add new bits to a bitfield
}
