% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1registry.R
\docType{class}
\name{registry-class}
\alias{registry-class}
\alias{registry}
\title{Bit registry class (S4) and methods}
\description{
A \code{registry} stores metadata and flag configuration of a bitfield.
}
\section{Slots}{

\describe{
\item{\code{width}}{\code{\link[=integer]{integerish(1)}}\cr how many bits is the bitfield wide.}

\item{\code{length}}{\code{\link[=integer]{integerish(1)}}\cr how many observations are encoded
in the bitfield.}

\item{\code{name}}{\code{\link[=character]{character(1)}}\cr short name of the bitfield.}

\item{\code{version}}{\code{\link[=character]{character(1)}}\cr automatically created version
tag of the bitfield. This consists of the package version, the version of R
and the date of creation of the bitfield.}

\item{\code{md5}}{\code{\link[=character]{character(1)}}\cr the MD5 checksum of the bitfield as
determined with \link[tools]{md5sum}.}

\item{\code{description}}{\code{\link[=character]{character(1)}}\cr longer description of the
bitfield.}

\item{\code{flags}}{\code{\link[=list]{list(.)}}\cr list of flags in the registry.}
}}

