% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1bitfield.R
\name{show,bitfield-method}
\alias{show,bitfield-method}
\title{Print geom in the console}
\usage{
\S4method{show}{bitfield}(object)
}
\arguments{
\item{object}{\link{\code{geom}}\cr object to \code{show}.}
}
\description{
Print geom in the console
}
