% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1bitfield.R
\name{show,bitfield-method}
\alias{show,bitfield-method}
\title{Print bitfield in the console (not working yet)}
\usage{
\S4method{show}{bitfield}(object)
}
\arguments{
\item{object}{\code{\link[=bitfield]{bitfield(1)}}\cr object to \code{show}.}
}
\description{
Print bitfield in the console (not working yet)
}
