% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1registry.R
\name{show,registry-method}
\alias{show,registry-method}
\title{Print registry in the console (not working yet)}
\usage{
\S4method{show}{registry}(object)
}
\arguments{
\item{object}{\code{\link[=registry]{registry(1)}}\cr object to \code{show}.}
}
\description{
Print registry in the console (not working yet)
}
